cleaned up some commented-out code, organized files in ecs
